make attachmenthandler and corefhandler as refiners
