+++ src/main/resources/eidos.conf
@@ -12,6 +12,7 @@ EidosSystem {
   masterRulesPath = ${EidosSystem.path}/grammars/master.yml
      taxonomyPath = ${EidosSystem.path}/grammars/taxonomy.yml
       hedgingPath = ${EidosSystem.path}/confidence/hedging.txt
+
 }
 
 filtering {
@@ -19,13 +20,6 @@ filtering {
 
     stopWordsPath = ${filtering.path}/stops.txt
   transparentPath = ${filtering.path}/transparent.txt
-}
-
-actions {
-  useCoref = true
-  corefType = "causalBasic"
-  taxonomyPath = ${EidosSystem.taxonomyPath}
-  useExpansion = true
 }
 
 apps {
@@ -74,24 +68,9 @@ ontologies {
   // Primary
   // Note that these first two are included via the build.sbt libraryDependencies on WorldModelers % Ontologies.
   wm               = ${ontologies.path}/wm_metadata.yml
-  wm_flattened     = ${ontologies.path}/wm_with_flattened_interventions_metadata.yml
+  wm_flattened     = ${ontologies.path}/wm_flat_metadata.yml
   wm_compositional = ${ontologies.path}/wm_compositional_metadata.yml
   interventions    = ${ontologies.path}/interventions_metadata.yml
-  // Legacy
-  un               = ${ontologies.path}/un_ontology.yml
-  props            = ${ontologies.path}/un_properties.yml
-  // Plugins
-  interventions    = ${ontologies.path}/interventions_metadata.yml
-  // Indicators
-  mitre12          = ${ontologies.path}/mitre12_indicators.yml
-  who              = ${ontologies.path}/who_ontology.yml
-  MaaS-model       = ${ontologies.path}/MaaS-model-ontology-edited.yaml
-  MaaS-parameter   = ${ontologies.path}/MaaS-parameter-ontology.yaml
-  MaaS-variable    = ${ontologies.path}/MaaS-variable-ontology.yaml
-  // Variables
-  icasa            = ${ontologies.path}/icasa.yml
-  // Other
-  mesh             = ${ontologies.path}/mesh_ontology.yml
 }
 
 geonorm {
@@ -102,14 +81,6 @@ geonorm {
 timenorm {
   useNeuralParser = false
   timeRegexPath = ${EidosSystem.path}/context/timenorm-regexes.txt
-}
-
-adjectiveGrounder {
-  useAdjectives = true
-  path = ${EidosSystem.path}/quantifierKB
-
-  domainParamKBPath = ${adjectiveGrounder.path}/domain_parameters.kb
-   quantifierKBPath = ${adjectiveGrounder.path}/gradable_adj_fullmodel.kb
 }
 
 // Context Finder (Time, Location)
