+++ src/main/resources/reference.conf
@@ -26,8 +26,6 @@ filtering {
 }
 
 actions {
-  useCoref = true
-  corefType = "causalBasic"
   taxonomyPath = ${EidosSystem.taxonomyPath}
   useExpansion = true
   expander = { include required("englishActionsExpander.conf") }
@@ -62,20 +60,27 @@ ontologies {
   ontologies = ["un", "wdi", "fao", "props", "mitre12", "who", "interventions"] // , "icasa", "mesh"]
   path = ${EidosSystem.path}/ontologies
 
-  // Legacy
-  un            = ${ontologies.path}/un_ontology.yml
-  props         = ${ontologies.path}/un_properties.yml
-  // Plugins
-  interventions = ${ontologies.path}/interventions_metadata.yml
-  // Indicators
-  mitre12       = ${ontologies.path}/mitre12_indicators.yml
-  who           = ${ontologies.path}/who_ontology.yml
-  // Variables
-  icasa         = ${ontologies.path}/icasa.yml
-  // Other
-  mesh          = ${ontologies.path}/mesh_ontology.yml
-  wdi           = ${ontologies.path}/wdi_ontology.yml
-  fao           = ${ontologies.path}/fao_variable_ontology.yml
+  // Primary
+    // Note that these first two are included via the build.sbt libraryDependencies on WorldModelers % Ontologies.
+    wm               = ${ontologies.path}/wm_metadata.yml
+    wm_flattened     = ${ontologies.path}/wm_with_flattened_interventions_metadata.yml
+    wm_compositional = ${ontologies.path}/wm_compositional_metadata.yml
+    interventions    = ${ontologies.path}/interventions_metadata.yml
+    // Legacy
+    un               = ${ontologies.path}/un_ontology.yml
+    props            = ${ontologies.path}/un_properties.yml
+    // Plugins
+    interventions    = ${ontologies.path}/interventions_metadata.yml
+    // Indicators
+    mitre12          = ${ontologies.path}/mitre12_indicators.yml
+    who              = ${ontologies.path}/who_ontology.yml
+    MaaS-model       = ${ontologies.path}/MaaS-model-ontology-edited.yaml
+    MaaS-parameter   = ${ontologies.path}/MaaS-parameter-ontology.yaml
+    MaaS-variable    = ${ontologies.path}/MaaS-variable-ontology.yaml
+    // Variables
+    icasa            = ${ontologies.path}/icasa.yml
+    // Other
+    mesh             = ${ontologies.path}/mesh_ontology.yml
 }
 
 ruleBasedEntityFinder {
@@ -110,4 +115,10 @@ adjectiveGrounder {
 
   domainParamKBPath = ${adjectiveGrounder.path}/domain_parameters.kb
    quantifierKBPath = ${adjectiveGrounder.path}/gradable_adj_fullmodel.kb
+}
+
+coref {
+    useCoref = true
+    corefType = "causalBasic"
+    taxonomyPath = ${EidosSystem.taxonomyPath}
 }
