+++ src/main/scala/org/clulab/wm/eidos/EidosActions.scala
@@ -11,7 +11,7 @@ import org.clulab.odin.EventMention
 import org.clulab.processors.Sentence
 import org.clulab.struct.Interval
 import org.clulab.wm.eidos.attachments._
-import org.clulab.wm.eidos.actions.{CorefHandler, MigrationHandler}
+import org.clulab.wm.eidos.actions.MigrationHandler
 import org.clulab.wm.eidos.attachments.CountModifier._
 import org.clulab.wm.eidos.attachments.CountUnit._
 import org.clulab.wm.eidos.expansion.Expander
@@ -22,7 +22,7 @@ import org.clulab.wm.eidos.utils.TagSet
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-import scala.collection.mutable.{Set => MutableSet}
+
 import scala.util.Try
 
 // 1) the signature for an action `(mentions: Seq[Mention], state: State): Seq[Mention]`
@@ -30,7 +30,7 @@ import scala.util.Try
 
 //TODO: need to add polarity flipping
 
-class EidosActions(val expansionHandler: Option[Expander], val coref: Option[CorefHandler]) extends Actions with LazyLogging {
+class EidosActions(val expansionHandler: Option[Expander]) extends Actions with LazyLogging {
   type Provenance = (String, Int, Int) // text, startOffset, endOffset
   type CountAndProvenance = (Double, Provenance)
   type CountUnitAndProvenanceOpt = (CountUnit.Value, Option[Provenance])
@@ -45,27 +45,16 @@ class EidosActions(val expansionHandler: Option[Expander], val coref: Option[Cor
   def globalAction(mentions: Seq[Mention], state: State): Seq[Mention] = {
     // Expand mentions, if enabled
     val expanded = expansionHandler.map(_.expand(mentions, state)).getOrElse(mentions)
-    // Merge attachments
-    val merged = mergeAttachments(expanded, state.updated(expanded))
-    // Keep only the most complete version of any given Mention
-//    /*val mostComplete =*/ keepMostCompleteEvents(merged, state.updated(merged))
-    // If the cause of an event is itself another event, replace it with the nested event's effect
-    // collect all effects from causal events
-    val (causal, nonCausal) = merged.partition(m => EidosSystem.CAG_EDGES.contains(m.label))
 
-    val assemble1 = createEventChain(causal, "effect", "cause")
-    // FIXME please
-    //val assemble2 = createEventChain(assemble1, "cause", "effect")
-    // in the sentence below we stitch together the sequence of cause->effect events
-    // but some expanded nounphrase remains, which shouldn't be displayed in the webapp
-    // In Kenya , the shortened length of the main growing season , due in part to a delayed onset of seasonal rainfall , coupled with long dry spells and below-average rainfall is resulting in below-average production prospects in large parts of the eastern , central , and southern Rift Valley .
-    val modifiedMentions = assemble1 ++ nonCausal
+    // Stitch together causal chains
+    val (causal, nonCausal) = expanded.partition(m => EidosSystem.CAG_EDGES.contains(m.label))
+    val assembled = createEventChain(causal, "effect", "cause")
 
-    // Basic coreference, if enabled
-    val afterResolving = coref.map(_.resolveCoref(modifiedMentions, state)).getOrElse(modifiedMentions)
+    // TODO: in the sentence below we stitch together the sequence of cause->effect events
+    //  but some expanded nounphrase remains, which shouldn't be displayed in the webapp
+    //    In Kenya , the shortened length of the main growing season , due in part to a delayed onset of seasonal rainfall , coupled with long dry spells and below-average rainfall is resulting in below-average production prospects in large parts of the eastern , central , and southern Rift Valley .
 
-    // I know I'm an unnecessary line of code, but I am useful for debugging and there are a couple of things left to debug...
-    afterResolving
+    assembled ++ nonCausal
   }
 
   protected def getCount(groupArg: Mention, numberArg: NumberArg): CountAndProvenance = {
@@ -344,96 +333,11 @@ class EidosActions(val expansionHandler: Option[Expander], val coref: Option[Cor
 
   def pass(mentions: Seq[Mention], state: State): Seq[Mention] = mentions
 
-  // Currently used as a GLOBAL ACTION in EidosSystem:
-  // Merge many Mentions of a single entity that have diff attachments, so that you have only one entity with
-  // all the attachments.  Also handles filtering of attachments of the same type whose triggers are substrings
-  // of each other.
-  def mergeAttachments(mentions: Seq[Mention], state: State): Seq[Mention] = {
 
-    // Get all the entity mentions for this span (i.e. all the "rainfall in Spain" mentions)
-    val (entities, nonentities) = mentions.partition(mention => mention matches "Entity")
-    val entitiesBySpan = entities.groupBy(entity => (entity.sentence, entity.tokenInterval, entity.label))
-    val mergedEntities = for {
-      (_, entities) <- entitiesBySpan
-      // These are now for the same span, so only one should win as the main one.
-      flattenedTriggeredAttachments = entities.flatMap(_.attachments.collect{ case a: TriggeredAttachment => a })
-      flattenedContextAttachments = entities.flatMap(_.attachments.collect{ case a: ContextAttachment => a })
-      filteredAttachments = filterAttachments(flattenedTriggeredAttachments)
-    }
-    yield {
-      val bestEntities =
-          if (filteredAttachments.nonEmpty) {
-            val bestAttachment = filteredAttachments.sorted.reverse.head
-            // Since head was used above and there could have been a tie, == should be used below
-            // The tie can be broken afterwards.
-            entities.filter(_.attachments.exists(_ == bestAttachment))
-          }
-          else
-            entities
-
-      MentionUtils.withOnlyAttachments(mostCompleteEventsKeeper.tieBreaker(bestEntities), filteredAttachments ++ flattenedContextAttachments)
-    }
-
-    val res = mostCompleteEventsKeeper.keepMostCompleteEvents(mergedEntities.toSeq ++ nonentities, state)
-    res
-  }
-
-  // Filter out substring attachments, then keep most complete.
-  def filterAttachments(attachments: Seq[TriggeredAttachment]): Seq[TriggeredAttachment] = {
-    attachments
-        // Perform first mapping based on class
-        .groupBy(_.getClass)
-        // Filter out substring attachments
-        .flatMap { case (_, attachments) => filterSubstringTriggers(attachments) }
-        // Next map based on both class and trigger.
-        .groupBy(attachment => (attachment.getClass, attachment.trigger))
-        // Now that substrings are filtered, keep only most complete of each class-trigger-combo.
-        .map { case (_, attachments) => filterMostComplete(attachments.toSeq) }
-        .toSeq
-  }
-
-  // Keep the most complete attachment here.
-  protected def filterMostComplete(attachments: Seq[TriggeredAttachment]): TriggeredAttachment =
-      attachments.maxBy(_.argumentSize)
-
-  // Filter out substring attachments.
-  protected def filterSubstringTriggers(attachments: Seq[TriggeredAttachment]): Seq[TriggeredAttachment] = {
-    val triggersKept = MutableSet[String]() // Cache triggers of itermediate results.
-
-    attachments
-        .sorted
-        .reverse
-        .filter { attachment =>
-          val trigger = attachment.trigger
-
-          if (!triggersKept.exists(_.contains(trigger))) {
-            triggersKept.add(trigger) // Add this trigger.
-            true // Keep the attachment.
-          }
-          else
-            false
-        }
-  }
-
-  // Get trigger from an attachment
-  protected def triggerOf(attachment: Attachment): String = {
-    attachment match {
-      case inc: Increase => inc.trigger
-      case dec: Decrease => dec.trigger
-      case quant: Quantification => quant.trigger
-      case _ => throw new UnsupportedClassVersionError()
-    }
-  }
 }
 
 object EidosActions extends Actions {
   lazy val logger: Logger = LoggerFactory.getLogger(this.getClass)
-
-  // Used for simplistic coreference identification
-  val COREF_DETERMINERS: Set[String] = Set("this", "that", "these", "those")
-
-  val ANTECEDENT: String = "antecedent"
-  val ANAPHOR: String = "anaphor"
 
   val BAD_NUMBER_PATTERN: Pattern = Pattern.compile("[~%,<>]")
 
@@ -446,12 +350,10 @@ object EidosActions extends Actions {
   val MAX_PATTERN: Pattern = Pattern.compile("""(at\s+most)|(less\s+than)|almost|under""")
 
   def fromConfig(config: Config, tagSet: TagSet): EidosActions = {
-    val useCoref: Boolean = config[Boolean]("useCoref")
-    val corefHandler = if (useCoref) Some(CorefHandler.fromConfig(config)) else None
 
     val useExpansion: Boolean = config[Boolean]("useExpansion")
     val expansionHandler = if (useExpansion) Some(Expander.fromConfig(config[Config]("expander"), tagSet)) else None
 
-    new EidosActions(expansionHandler, corefHandler)
+    new EidosActions(expansionHandler)
   }
 }
